Added running total service
