+++ MyMoney.Core/Interfaces/Service/IIncomeService.cs
@@ -13,6 +13,5 @@ namespace MyMoney.Core.Interfaces.Service
       IIncome Find(long incomeId);
       IList<IIncome> From(DateTime start, int count);
       IList<IIncome> Between(DateTime start, DateTime end);
-      IList<RunningTotal> RunningTotal(decimal initialTotal, DateTime start, DateTime end);
    }
 }
