+++ MyMoney.Core/Interfaces/Service/IRunningTotalService.cs
@@ -1,0 +1,11 @@
+﻿using MyMoney.Core.Data;
+using System;
+using System.Collections.Generic;
+
+namespace MyMoney.Core.Interfaces.Service
+{
+   public interface IRunningTotalService
+   {
+      IList<RunningTotal> RunningTotal(decimal initialTotal, DateTime start, DateTime end);
+   }
+}
